Move wallets into top level directory wallet/
