+++ components/wallet-card.js
@@ -3,7 +3,7 @@ import styles from '@/styles/wallet.module.css'
 import Plug from '@/svgs/plug.svg'
 import Gear from '@/svgs/settings-5-fill.svg'
 import Link from 'next/link'
-import { Status } from './wallet'
+import { Status } from '../wallets'
 
 export default function WalletCard ({ wallet }) {
   const { card: { title, badges } } = wallet
