+++ pages/settings/wallets/index.js
@@ -2,7 +2,7 @@ import { getGetServerSideProps } from '@/api/ssrApollo'
 import Layout from '@/components/layout'
 import styles from '@/styles/wallet.module.css'
 import Link from 'next/link'
-import { useWallets, walletPrioritySort } from '@/components/wallet'
+import { useWallets, walletPrioritySort } from 'wallets'
 import { useEffect, useState } from 'react'
 import dynamic from 'next/dynamic'
