+++ wallets/index.js
@@ -5,17 +5,17 @@ import { useWalletLogger } from '@/components/wallet-logger'
 import { SSR } from '@/lib/constants'
 import { bolt11Tags } from '@/lib/bolt11'
 
-import * as lnbits from '@/components/wallet/lnbits'
-import * as nwc from '@/components/wallet/nwc'
-import * as lnc from '@/components/wallet/lnc'
-import * as lnd from '@/components/wallet/lnd'
-import * as lnAddr from '@/components/wallet/lightning-address'
-import * as cln from '@/components/wallet/cln'
+import * as lnbits from 'wallets/lnbits'
+import * as nwc from 'wallets/nwc'
+import * as lnc from 'wallets/lnc'
+import * as lnd from 'wallets/lnd'
+import * as lnAddr from 'wallets/lightning-address'
+import * as cln from 'wallets/cln'
 import { gql, useApolloClient, useQuery } from '@apollo/client'
 import { REMOVE_WALLET, WALLET_BY_TYPE } from '@/fragments/wallet'
 import { autowithdrawInitial } from '@/components/autowithdraw-shared'
 import { useShowModal } from '@/components/modal'
-import { useToast } from '../toast'
+import { useToast } from '../components/toast'
 
 // wallet definitions
 export const WALLET_DEFS = [lnbits, nwc, lnc, lnd, lnAddr, cln]
